Org unit sibling display sort order
Adds a new "sibling_order" column to actor.org_unit which specifies the
order in which an org unit should be sorted as compared to its sibling
org units in org unit trees.

This commit adds the IDL/DB components.  It also updates the tpac and
the generic get_org_tree API.  (Note, a follow-up commit will be pushed
to remove some deprecated / reduntant org tree retrievals).

Note that the sibling_order values do not have to be unique within each
set of siblings.  The org unit name is still used as a tie-breaker sort,
so if the sort order is left unset (i.e. all values are 0), org units
will fall back to name-only sorting.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>