AutoSuggest: don't force-scroll browser; that's irritating
By default, Dijits and things that inherit from Dijits (at least in
Dojo 1.3.3 that we're still using), have a true value for an attribute
called scrollOnFocus.  This makes your browser jump to ensure that the
widget that just received focus is in view.

You don't always want this though.  Try expanding the extras on the
record detail page of the TPAC without this commit, but with AutoSuggest
enabled, for example.

This commit sets the scrollOnFocus attribute to false for the
AutoSuggest widget.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>