Staff client: add a back-to-search-results button in the chrome that ...
... wraps the OPAC (whether JSPAC or TPAC).

You don't really need this when in OPAC view, because the OPAC page
itself will have a link that does the same thing.  However, if you're
browsing through records in a search result set one at a time in some
other view, say the MARC view, then this button can come in handy.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Michael Peters <mrpeters@library.in.gov>
Signed-off-by: Mike Rylander <mrylander@gmail.com>