LP#707167 Make open-ils.actor.usrname.exists not ignore deleted users
Currently, open-ils.actor.usrname.exists filters out deleted users.  This
results in a user recieving indication as to why a user update
failed, when the reason is the username they're attempting to use
is in use, but is "deleted" in the database.

This branch makes open-ils.actor.usrname.exists look at deleted users
since they still prevent re-use of the delted user's username.

More information is in the original bug, https://bugs.launchpad.net/evergreen/+bug/707167.

Signed-off-by: Michael Peters <mrpeters@library.in.gov>
Signed-off-by: Mike Rylander <mrylander@gmail.com>