Tpac - add additional subfields for title in search results
Originally, titles in Tpac only showed based on tag 245 subfield a.
This sometimes led to difficulty identifying specific records from
each other if it was part of a set or had additional title information
in other subfields.

Modified default title function to include the following:

subfield b = remainder of title
subfield n = number of part/section of a work
subfield p = name of part/section of a work

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Dan Scott <dan@coffeecode.net>