TPAC: Prevent OU-invisible copies from being displayed to public
Per LP# 959577, "We've found that copies from OU's with OPAC visible set
to false are displaying on the record details page of tpac."

This fixes that by including the pertinent clause in mk_copy_query()'s
version of what in-db unapi does. Some day we will unite under the same
logic and be stronger!

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>