Custom Org Tree : DB and IDL
Support for building custom org unit hierarchies for display purposes.
Initially, this is meant to support custom OPAC trees, but the design
leaves from for other types of trees.

The only restrictions to custom org trees is that no org unit appear
more once and that they in fact be tree-shaped.  Otherwise, any node
can be the parent/child of any other node, regardless of ou_type, etc.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>