Custom Org Tree : tpac
If configured and active, the tpac will now use the custom 'opac' tree
for the org unit selectors.  This will be true in both the public
catalog and the staff client catalog.  The staff client catalog will
continue to display hidden org units.

If the opac.org_unit.non_inheritied_visibility global flag is set, the
public catalog will display child nodes of hidden org units in the
custom tree.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>