Teach Fieldmapper (Perl) to record and expose field datatype
We record most everything from the IDL, but until now we skipped datatype
because it's generally not useful in perl.  But, there might be uses.  So
now we record it, and expose it through the class and instance FieldDatatype
method.  Also added is a more general FieldInfo method, which returns a hash
containing field attributes, including but not limited to:

 * virtual  [1|0]
 * required [1|0]
 * position [array position for internal implementation]
 * datatype [bool|float|id|int|interval|link|money|number|org_unit|text|timestamp]
 * validate [regexp]

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>