AutoFieldWidget single-object cache repairs
Caching single objects linked off the context object was failing in most
cases as it was caching under the display name instead of the true value
(pkey) for the linked object.  This fixes that by always caching the
linked object by its true value, so that future cache lookups will be
looking in the right place.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>