ACQ Upload import-copy option
With the addition of the new Vandelay features in ACQ, the "Load bibs
and Items" option was replaced with Vandeley import options.  In the
context of ACQ, however, Vandelay is only concerned with importing bibs
and not copies.  Subsequently, the user lost the ability to both create
catalog records and catalog copies in one step during ACQ upload.

The patch remedies that by providing a new "Load Items for Imported
Records" option within the upload page.  When enabled, all successfully
loaded ACQ copies will also be imported as real/catalog copies.

Note that it's only added to the upload page, because all other ACQ
interfaces assume that both bibs and copies should be created (e.g.
during PO activation).

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>