TPAC: Suppress copy counts for libraries that own zero copies
Rather than slavishly showing the copy counts for libraries even when
the library has no copies to show, avoid wasting the precious display
space and suppress the output.

This was a problem for both search scope libraries and preferred library
in the search results, and for just preferred library in record details.
This commit makes the behaviour consistent, in favour of suppressing the
display of copy counts when there's nothing to display.

Signed-off-by: Dan Scott <dan@coffeecode.net>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>