Hold targeter cancellation action/trigger repairs
When a hold/reservation is cancelled via the targeter and passed off to
action/trigger for possible event creation, ensure that all fields on
the hold are in sync w/ the master db before the hand-off.  Otherwise,
critical values, like cancel_time may be null, which can cause A/T
event creation to fail.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>