TPAC: Centralized preferred library indicator
Create an explicit "Preferred library: Foobar" entry at the top of the
page (on the same line as pagination information), and include a subtle
checkmark to enable users to quickly jump to their "Search preferences"
settings and change their preferred library.

This enables us to remove the "(Preferred library)" note in the copy
counts for search results.

At the same time, move from a table display (for a single row? what?) to
a simple div/span layout for the search results.

Signed-off-by: Dan Scott <dan@coffeecode.net>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>