New pull list interface taking advantage of flattener for speed,
and advanced sorting.  For now, access it by the "Simplifed Pull List"
button along the bottom edge of the existing holds pull list interface
(but I think when/if this thing is widely accepted, it should replace
the existing interface outright).

With thanks to Mike Peters for testing an early version.  Now including
some updates requested by Thomas Berezansky.  Specifically, the
queue_position column and its relatives fthat come from the same
calculation were removed, as they're [very] expensive in computing time to
produce and generally useless to pull lists.

One exception to the characterization of those fields as "useless" is the
"number of potential copies" column, which we should add back later
assuming we can find a fast way to do it rather than the previous ways,
which were slow.

Thanks to Bill Erickson for helping fix my issues in making auto-generated
columns coöperate with the column picker (his changes are squashed into
this).  I think it's finally right.

Now with release notes.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>