FlattenerGrid (re)honors IDL column labels as default
Linking the GridColumnPicker to FlattenerGrid killed the ability for the
FlattenerGrid to fall back to the IDL column label when no label was
defined in the markup.  This change adds a reloadStructure operation to
GridColumnPicker.  FlattenerGrid uses this to pass the updated column
labels (and any other changes) to gridcolumnpicker, which then updates
the grid display to match.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>