TPAC: Show deleted record warning for deleted records
At the cost of an extra database query per record displayed, we can
properly activate the "This record has been deleted..." message. Prior
to this commit, the message was embedded on every page but hidden via
CSS - which isn't great for search engines or source-reading types.

Now we only include & display the message if it has, in fact, been
deleted.

Signed-off-by: Dan Scott <dan@coffeecode.net>
Signed-off-by: Ben Shum <bshum@biblio.org>