TPAC: Protect against requests for non-existent records
If a record request is received for a record that does not exist (say,
if 100 records have been loaded and someone submits a request for
/eg/opac/record/399), rather than dying with a server error, return
immediately.

Eventually we'll want to provide an explicit "record was not
found error" in the TPAC record details; for now the TPAC just displays
zero bib detail.

Signed-off-by: Dan Scott <dan@coffeecode.net>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>