Bug 957453: Fix error in database seed data
The use of an explicit ID in creating a new
action_trigger.event_definition row was followed by the use of the
CURRVAL() function to grab the most recently issued value for
the action_trigger.event_defintion_id_seq sequence - which happened to
be 100, due to a previously issued SETVAL() on the sequence.

This resulted in references to non-existent event definitions and thus
an error in creating the stock database schema.

By using an explicit ID for the dependent row, we avoid this error.

Signed-off-by: Dan Scott <dan@coffeecode.net>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>