LP#845096 - coded_value_map_id_seq doesn't have a proper last_value
Thanks to Kathy Lussier for the original report.  Kathy discovered
that users were unable to add new Coded Value Maps via the staff client.

With some log digging, I discovered the culprit was a rash of
primary key errors.

With help from Dan Scott, I was able to rectify this via a quick bit
of SQL which properly sets the last_value in the sequence.

Thanks, Kathy and Dan!

Signed-off-by: Michael Peters <mrpeters@library.in.gov>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>