add more penalty related CSS hooks
PATRON_HAS_CUSTOM_PENALTY (if the penalty is not a stock penalty, based on id)
PATRON_HAS_STAFF_ALERT (if the staff_alert boolean on the penalty is true)
PATRON_HAS_BLOCK (if there's anything in the penalty's block list)
PATRON_HAS_CIRC_BLOCK (if CIRC is in the block list)
PATRON_HAS_RENEW_BLOCK (if RENEW is in the block list)
PATRON_HAS_HOLD_BLOCK (if HOLD is in the block list)
PATRON_HAS_CAPTURE_BLOCK (if CAPTURE is in the block list)
PATRON_HAS_FULFILL_BLOCK (if FULFILL is in the block list)

also adds stock CSS for PATRON_HAS_STAFF_ALERT, a blue border around the patron
name.

In the color hierarchy, this comes after the normal Green/Lime status, but gets
trumped by any other penalty that sets a colored border.

So, for example, if a patron has no penalties, and you go to Messages and add an
"Alert" note, that equates to the penalty ALERT_NOTE, which has a staff_alert
setting of True.  On a refresh, the patron will then have a blue border around
their name.  If you then create a bill on that patron for $100, triggering the
PATRON_EXCEEDS_FINES penalty, then the patron will gain the normal purple or
lavender border, even though the penalty also has a staff_alert setting of True
and triggers PATRON_HAS_STAFF_ALERT CSS.

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>