TPac: Advanced Search Config
Move advanced search limiter config to config.tt2.

This should, in theory, make it much easier to customize the fields shown
for searching in TPac.

Also adds the VR Format Coded Value Maps so that we can populate that
limiter with them.

And removes dojo.css from being loaded - That includes reset stuff that
messes with our CSS but doesn't seem to give us any benefit.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>