AutoSuggest: Escape ampersands properly
This fixes an issue reported by Yamil Suarez.

If you had relatively technical users composing searches with
QueryParser syntax, or if they were just typing ampersands for any
other reason, AutoSuggest would behave as if the ampersand marked the end
of user input.

This is fixed by applying the correct URI-encoding function for the
situation.

To be clear, QueryParser syntax does not actually affect suggestions;
such syntax is ignored.  AutoSuggest is not search.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Dan Scott <dan@coffeecode.net>