Consensus was reached to remove editor-specific globs from .gitignore.
In a highly overengineered discussion out of scope with the actual breadth
of this very minor issue, everyone has come to the conclusion that
editor-specific globs don't really belong in the project's .gitignore,
given that git permits a global-level ignore file for users.

Instead, a HACKING file is herein added, with its first entry to explain
how developers can set the own global gitignore.

Signed-off-by: Bradley M. Kuhn <bkuhn@ebb.org>
Signed-off-by: Dan Scott <dan@coffeecode.net>