Merge the Evergreen docs into the code repository
Rather than maintaining two separate repositories, with dual maintenance
of what are essentially single-sourced files in the release notes and
the install instructions, make better use of the docs/ subfolder of the
Evergreen repository. If this crazy scheme works, DIG can commit to the
docs/ subfolder, and devs can commit there too.

Signed-off-by: Dan Scott <dscott@laurentian.ca>