Always display the hold type indicator
We should be showing the type of hold on any pull-list or hold detail
interface we build.  The primary use case is to alert staff when an
alternate identical copy of the same title will (very likely) be usable
by a hold, and (as in the case of C-like holds) it will not.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>