Allow more granular overrides
This adds a new parameter to backend override functions. In most cases the
system will act like a hash with all set to 1 was passed in when nothing was
passed in by the caller.

The new parameter should be a hash containing zero or more of the following
options:

all - If true (1) override all events
events - Should be an array of event textcodes to override

For hold placement the "all" option is ignored for possibility checks, but
is checked for other events like HOLD_EXISTS.

Also included for hold placement is to always run possibility checks for
the test and create batch method.

Overall, the default is to act like things did previously.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>