README: Backport changes to README from master (27320c83)
The OpenSRF prereq had already been bumped to 2.1 in the rel_2_2 branch,
which was unexpected as the same change had not occurred in master.
However, we can apply the other diffs from 27320c83 here:

Use a title that fits into the DIG documentation more smoothly, and
be consistent about sentence-style capitalization of sub-headings.

Signed-off-by: Dan Scott <dan@coffeecode.net>