Remove unsafe thesaurus/control-set mapping
Evergreen 2.2 and beyond includes new Authority Control Set functionality
intended to allow the definition of locally defined authority control
headings.  The primary mechanism for determining which Control Set is used
by any given authority record is to look at the thesaurus fixed field
(mnemonic: Subj), which currently must be uniquely linked to a Control Set.
Failing that (in the case of no thesaurus code at all), there is a backup
mechanism: if the thesaurus value for the authority record is missing or not
used by any configured Control Set then Evergreen will look at the first Main
Entry (1xx) field in the record, and find a control set which uses this field.
This is, obviously, not perfect, but for non-overlapping authority field
schemes, it works very well.

However, the seed data for Evergreen currently includes the "no attempt to
code" value of "|" in the set of thesauri that point to the LoC Control Set.
This value ("|") is also the default value for the Subj fixed field.  This
means that the fallback mechanism is very likely never reached, and anything
short of fully specifying an appropriate thesaurus code for any non-LoC
authority record will run afoul of a similar outcome as the bib-data loss
situation described in bug #983487.

We can avoid this situation in most cases by /not/ tying the "|" thesaurus
value to the LoC control set.

This commit accomplishes this goal by allowing the Control Set for a thesaurus
to be NULL, removing the Control Set for "|" and removing the Control Set
recorded for records with "|" as the thesaurus value, thereby forcing Main
Entry inspection on any use of the authority record where Control Set is
required.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>