Z39.50 attributes management interface
From the release notes:

There is a new interface for managing Z39.50 attributes on a Z39.50
source.  The interface is linked from each source name in the Z39.50
Source administrative interface.

Cloning
~~~~~~

In addition to attribute creation, deletion, and editing, it's also
possible to clone a set of attributes from one source into another.
When cloning, any attributes present in the cloned source that are
not present in the destination source are copied into the destination
source.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Dan Scott <dan@coffeecode.net>